Prevent hidden prompt-frame rows from seeding stale checkpoints
The branch had already absorbed most review-driven fixes, but rebasing onto current dev still left one real correctness gap: initial summary checkpoint materialization could trigger on raw overflow even when hidden internal rows meant the visible window had not overflowed yet. This patch switches the seed path to require visible overflow, carries forward the latest durable-flush test hardening from the live PR branch, aligns the delegate child profile fragment with advisory authority, reuses the shared sqlite test lock, and hardens benchmark/report helpers so missing prompt-frame hashes or token totals do not silently look healthy.

Constraint: The PR branch moved during review and had to be rebased onto current dev before final fixes
Constraint: Local tmp-target verification exhausted disk space, so validation moved to a shared persistent cargo target cache
Rejected: Keep raw-turn initial checkpoint seeding | hidden internal rows still materialize stale summary boundaries
Rejected: Ignore the newer durable-flush test follow-up commit | would overwrite live branch work and keep the snapshot race unaddressed
Confidence: medium
Scope-risk: moderate
Directive: Summary checkpoint seeding must prove visible overflow, not just raw session-turn overflow
Tested: cargo fmt --all -- --check; cargo check -p loongclaw-app --offline; cargo check -p loongclaw-bench --offline; cargo check -p loongclaw --offline
Not-tested: Full local test matrix; CI will provide the full cross-feature and runtime matrix
